+++ packages/uniforms-bridge-zod/src/register.ts
@@ -1,0 +1,11 @@
+import { filterDOMProps } from 'uniforms';
+
+// There's no possibility to retrieve them at runtime.
+declare module 'uniforms' {
+  interface FilterDOMProps {
+    minCount: never;
+    maxCount: never;
+  }
+}
+
+filterDOMProps.register('minCount', 'maxCount');