+++ packages/uniforms-unstyled/src/AutoFields.tsx
@@ -1,4 +1,4 @@
-import { ComponentType, createElement } from 'react';
+import { ComponentType, Fragment, createElement } from 'react';
 import { useForm } from 'uniforms';
 
 import AutoField from './AutoField';
@@ -10,7 +10,7 @@ export type AutoFieldsProps = {
 };
 
 export default function AutoFields({
-  element = 'div',
+  element = Fragment,
   fields,
   omitFields = [],
   ...props